requirements.txt
streamlit
streamlit-chat
openai
python-dotenv
tiktoken
langchain
docx2txt
pypdf
chromadb